Вынесение контроллера
